Change rating to float, implement RecommendationService

+++ src/main/java/com/moviesandchill/recommendationservice/pojo/UserFilmRating.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class UserFilmRating {
     private long userId;
     private long filmId;
-    private int rating;
+    private float rating;
 }
